example_project/requirements.txt
# Base
wsgiref==0.1.2

# Django base
svn+http://code.djangoproject.com/svn/django/trunk@15481#egg=Django

# Debug
django-debug-toolbar==0.8.4

# Apps
django-netauth
